CS Client/requirements.txt
# Core Dependencies
pytesseract>=0.3.0
Pillow>=9.0.0
flask>=2.0.0
flask-cors>=4.0.0  # For handling cross-origin requests